Refactor no BattlefieldSetup: função genérica pra cuidar do loop. Passagem das operações de trim pra uma stringutils. AskForNumberOfCharactersInPlayerTeam já começa a usar o loop genérico.
