+++ BattlefieldSetup.cpp
@@ -1,33 +1,28 @@
 #include "BattlefieldSetup.h"
 #include <iostream>
 #include <string>
-#include <algorithm>
+#include "stringutils.h"
 #include <functional>
-#include <cctype>
-#include <locale>
-
 using namespace std;
-//Obtido de https://stackoverflow.com/questions/216823/how-to-trim-an-stdstring, pq a stl n tem
-// trim de string por default e eu quero trimar a string.
-//  
-// trim from start
-std::string& ltrim(std::string& s) {
-    s.erase(s.begin(), std::find_if(s.begin(), s.end(),
-        std::not1(std::ptr_fun<int, int>(std::isspace))));
-    return s;
+/// <summary>
+/// Função genérica pra fazer o loop de pedir resultado.
+/// </summary>
+/// <typeparam name="ReturnValue">O tipo do que o prompt retorna</typeparam>
+/// <param name="bar">A função do prompt. Seu bool& deve retornar true se o prompt foi preeenchido c/sucesso
+/// ou false caso contrário pq ele é o usado pra continuar o loop ou não</param>
+/// <returns></returns>
+template<typename ReturnValue>
+ReturnValue foo(ReturnValue(*bar)(bool&)) {
+	bool validChoice = false;
+	while (!validChoice) {
+		ReturnValue result = bar(validChoice);
+		if (validChoice)
+			return result;
+	}
 }
 
-// trim from end
-std::string& rtrim(std::string& s) {
-    s.erase(std::find_if(s.rbegin(), s.rend(),
-        std::not1(std::ptr_fun<int, int>(std::isspace))).base(), s.end());
-    return s;
-}
+int AskForNumberOfCharactersInPlayerTeam();
 
-// trim from both ends
-std::string& trim(std::string& s) {
-    return ltrim(rtrim(s));
-}
 /// <summary>
 /// Encapsula o processo de pedir a classe ao player.
 /// </summary>
@@ -44,91 +39,114 @@ int ChooseLines();
 /// <returns></returns>
 int ChooseRows();
 
+
 GameSetupParameters AskForParameters() {
-    GameSetupParameters response;
-    response.PlayerClassId = ChooseClass();
-    response.GridLines = ChooseLines();
-    response.GridRows = ChooseRows();
-    return response;
+	GameSetupParameters response;
+	int numberOfCharactersInPlayersTeam = AskForNumberOfCharactersInPlayerTeam();
+	response.PlayerClassId = ChooseClass();
+	response.GridLines = ChooseLines();
+	response.GridRows = ChooseRows();
+	return response;
 }
+
 bool AskIfWantToPlayAgain()
 {
-    bool validChoice = false;
-    bool response = false;
-    while (!validChoice) {
-        cout << "Game is over. Would you like to play again? [y/n] " << endl;
-        string choiceAsStr;
-        std::getline(std::cin, choiceAsStr);
-        choiceAsStr = trim(choiceAsStr);
-        if (choiceAsStr == "y" || choiceAsStr == "Y") {
-            response = true;
-            validChoice = true;
-        }
-        else if (choiceAsStr == "n" || choiceAsStr == "N")
-        {
-            response = false;
-            validChoice = true;
-        }
-        else {
-            cout << "Invalid option:" << choiceAsStr << endl;
-        }
-    }
-    return response;
+	bool validChoice = false;
+	bool response = false;
+	while (!validChoice) {
+		cout << "Game is over. Would you like to play again? [y/n] " << endl;
+		string choiceAsStr;
+		std::getline(std::cin, choiceAsStr);
+		choiceAsStr = trim(choiceAsStr);
+		if (choiceAsStr == "y" || choiceAsStr == "Y") {
+			response = true;
+			validChoice = true;
+		}
+		else if (choiceAsStr == "n" || choiceAsStr == "N")
+		{
+			response = false;
+			validChoice = true;
+		}
+		else {
+			cout << "Invalid option:" << choiceAsStr << endl;
+		}
+	}
+	return response;
 }
 int ReadPositiveInt(const std::string& text) {
-    bool validChoice = false;
-    int result = -1;
-    while (validChoice == false) {
-        cout << text << endl;
-        std::string choiceAsStr;
-        std::getline(std::cin, choiceAsStr);
-        choiceAsStr = trim(choiceAsStr);
-        try {
-            int choice = std::stoi(choiceAsStr);
-            if (choice > 0) {
-                result = choice;
-                validChoice = true;
-            }
-            else {
-                throw std::invalid_argument("must be above 0");
-            }
-        }
-        catch (std::invalid_argument const& ex) {
-            cout << "Invalid Value..."<<ex.what() << endl;
-            validChoice = false;
-        }
-    }
-    return result;
+	bool validChoice = false;
+	int result = -1;
+	while (validChoice == false) {
+		cout << text << endl;
+		std::string choiceAsStr;
+		std::getline(std::cin, choiceAsStr);
+		choiceAsStr = trim(choiceAsStr);
+		try {
+			int choice = std::stoi(choiceAsStr);
+			if (choice > 0) {
+				result = choice;
+				validChoice = true;
+			}
+			else {
+				throw std::invalid_argument("must be above 0");
+			}
+		}
+		catch (std::invalid_argument const& ex) {
+			cout << "Invalid Value..." << ex.what() << endl;
+			validChoice = false;
+		}
+	}
+	return result;
 
 }
 int ChooseLines()
 {
-    return ReadPositiveInt("How many lines?");
+	return ReadPositiveInt("How many lines?");
 }
 int ChooseRows() {
-    return ReadPositiveInt("How many rows?");
+	return ReadPositiveInt("How many rows?");
 }
 CharacterClass ChooseClass()
 {
-    bool classChoiceIsOk = false;
-    CharacterClass result = PALADIN;
-    while (classChoiceIsOk == false) {
-        cout << "Choose Between One of this Classes:" << endl;
-        cout << "[1] Paladin, [2] Warrior, [3] Cleric, [4] Archer " << endl;
-        std::string choiceAsStr;
-        std::getline(std::cin, choiceAsStr);
-        choiceAsStr = trim(choiceAsStr);
-        try {
-            CharacterClass choice = std::stoi(choiceAsStr);
-            classChoiceIsOk = (choice == PALADIN || choice == WARRIOR || choice == CLERIC || choice == ARCHER);
-            if(!classChoiceIsOk)
-                throw std::invalid_argument("out of range");
-            result = choice;
-        }
-        catch (std::invalid_argument const& ex) {
-            cout << "Invalid value..." << ex.what() << endl;
-            classChoiceIsOk = false;
-        }
-    }
-    return result;
+	bool classChoiceIsOk = false;
+	CharacterClass result = PALADIN;
+	while (classChoiceIsOk == false) {
+		cout << "Choose Between One of this Classes:" << endl;
+		cout << "[1] Paladin, [2] Warrior, [3] Cleric, [4] Archer " << endl;
+		std::string choiceAsStr;
+		std::getline(std::cin, choiceAsStr);
+		choiceAsStr = trim(choiceAsStr);
+		try {
+			CharacterClass choice = std::stoi(choiceAsStr);
+			classChoiceIsOk = (choice == PALADIN || choice == WARRIOR || choice == CLERIC || choice == ARCHER);
+			if (!classChoiceIsOk)
+				throw std::invalid_argument("out of range");
+			result = choice;
+		}
+		catch (std::invalid_argument const& ex) {
+			cout << "Invalid value..." << ex.what() << endl;
+			classChoiceIsOk = false;
+		}
+	}
+	return result;
+}
+int AskForNumberOfCharactersInPlayerTeam() {
+	return foo<int>([](bool& validInput)->int
+		{
+			try {
+				cout << "How many characters on your team? " << endl;
+				string choiceAsStr;
+				std::getline(std::cin, choiceAsStr);
+				int choiceAsNumber = stoi(choiceAsStr);
+				if (choiceAsNumber <= 0)
+					throw std::invalid_argument("Must have at least one character on the team");
+				validInput = true;
+				return choiceAsNumber;
+			}
+			catch (std::invalid_argument const& ex) {
+				cout << "Invalid Value..." << ex.what() << endl;
+				validInput = false;
+				return -1;
+			}
+		});
 }
